- Fix add workspace mgmt api call - Make calls to mgmt api only in production env.

+++ services/users/server/api/controllers/users.controller.ts
@@ -161,22 +161,24 @@ export class UsersControllers {
             }
 
             // Send user to the mgmt portal
-            let userMgmt = {
-                _id: user._id,
-                active: user.active,
-                email: user.email,
-                password: user.password,
-                first_name: user.first_name,
-                last_name: user.last_name,
-                _workspace: user._workspace,
-                environment: "PROD", // TODO
-                created_date: user.created_date
-            }
+            if (process.env.NODE_ENV !== 'production') {
+                let userMgmt = {
+                    _id: user._id,
+                    active: user.active,
+                    email: user.email,
+                    password: user.password,
+                    first_name: user.first_name,
+                    last_name: user.last_name,
+                    _workspace: user._workspace,
+                    environment: "PROD", // TODO
+                    created_date: user.created_date
+                }
 
-            http.put(`${process.env.MANAGEMENT_URL}/api/user/${userMgmt._id}/update`, {
-                API_KEY: process.env.MANAGEMENT_API_KEY,
-                userData: userMgmt
-            });
+                http.put(`${process.env.MANAGEMENT_URL}/api/user/${userMgmt._id}/update`, {
+                    API_KEY: process.env.MANAGEMENT_API_KEY,
+                    userData: userMgmt
+                });
+            }
 
             // Send status 200 response
             return res.status(200).json({
@@ -222,43 +224,45 @@ export class UsersControllers {
             await userBY.save();
 
             // Send new workspace to the mgmt portal
-            // Count all the users present inside the workspace
-            const usersCount: number = await User.find({ $and: [
-                { active: true },
-                { _workspace: workspaceId }
-            ]}).countDocuments();
+            if (process.env.NODE_ENV !== 'production') {
+                // Count all the users present inside the workspace
+                const usersCount: number = await User.find({ $and: [
+                    { active: true },
+                    { _workspace: workspaceId }
+                ]}).countDocuments();
 
-            // Count all the groups present inside the workspace
-            const groupsCount: number = await Group.find({ $and: [
-                { group_name: { $ne: 'personal' } },
-                { _workspace: workspaceId }
-            ]}).countDocuments();
+                // Count all the groups present inside the workspace
+                const groupsCount: number = await Group.find({ $and: [
+                    { group_name: { $ne: 'personal' } },
+                    { _workspace: workspaceId }
+                ]}).countDocuments();
 
-            let workspaceMgmt = {
-                _id: workspaceId,
-                company_name: workspace.company_name,
-                workspace_name: workspace.workspace_name,
-                owner_email: workspace.owner_email,
-                owner_first_name: workspace.owner_first_name,
-                owner_last_name: workspace.owner_last_name,
-                _owner_remote_id: workspace._owner,
-                environment: "PROD", // TODO
-                num_members: usersCount,
-                num_invited_users: workspace.invited_users.length,
-                num_groups: groupsCount,
-                created_date: workspace.created_date,
-                billing: {
-                    subscription_id: workspace.billing.subscription_id,
-                    current_period_end: workspace.billing.current_period_end,
-                    scheduled_cancellation: workspace.billing.scheduled_cancellation,
-                    quantity: usersCount
+                let workspaceMgmt = {
+                    _id: workspaceId,
+                    company_name: workspace.company_name,
+                    workspace_name: workspace.workspace_name,
+                    owner_email: workspace.owner_email,
+                    owner_first_name: workspace.owner_first_name,
+                    owner_last_name: workspace.owner_last_name,
+                    _owner_remote_id: workspace._owner,
+                    environment: "PROD", // TODO
+                    num_members: usersCount,
+                    num_invited_users: workspace.invited_users.length,
+                    num_groups: groupsCount,
+                    created_date: workspace.created_date,
+                    billing: {
+                        subscription_id: workspace.billing.subscription_id,
+                        current_period_end: workspace.billing.current_period_end,
+                        scheduled_cancellation: workspace.billing.scheduled_cancellation,
+                        quantity: usersCount
+                    }
                 }
-            }
 
-            http.put(`${process.env.MANAGEMENT_URL}/api/workspace/${workspace._id}/update`, {
-                API_KEY: process.env.MANAGEMENT_API_KEY,
-                workspaceData: workspaceMgmt
-            });
+                http.put(`${process.env.MANAGEMENT_URL}/api/workspace/${workspace._id}/update`, {
+                    API_KEY: process.env.MANAGEMENT_API_KEY,
+                    workspaceData: workspaceMgmt
+                });
+            }
             
             // Send status 200 response
             return res.status(200).json({
@@ -659,43 +663,45 @@ export class UsersControllers {
         });
 
         // Send new workspace to the mgmt portal
-        // Count all the groups present inside the workspace
-        const groupsCount: number = await Group.find({ $and: [
-            { group_name: { $ne: 'personal' } },
-            { _workspace: workspace._id }
-        ]}).countDocuments();
+        if (process.env.NODE_ENV !== 'production') {
+            // Count all the groups present inside the workspace
+            const groupsCount: number = await Group.find({ $and: [
+                { group_name: { $ne: 'personal' } },
+                { _workspace: workspace._id }
+            ]}).countDocuments();
 
-        let workspaceMgmt = {
-            _id: workspace._id,
-            company_name: workspace.company_name,
-            workspace_name: workspace.workspace_name,
-            owner_email: workspace.owner_email,
-            owner_first_name: workspace.owner_first_name,
-            owner_last_name: workspace.owner_last_name,
-            _owner_remote_id: workspace._owner,
-            environment: "PROD", // TODO
-            num_members: usersCount,
-            num_invited_users: workspace.invited_users.length,
-            num_groups: groupsCount,
-            created_date: workspace.created_date,
-            billing: {
-                subscription_id: workspace.billing.subscription_id,
-                current_period_end: workspace.billing.current_period_end,
-                scheduled_cancellation: workspace.billing.scheduled_cancellation,
-                quantity: usersCount
+            let workspaceMgmt = {
+                _id: workspace._id,
+                company_name: workspace.company_name,
+                workspace_name: workspace.workspace_name,
+                owner_email: workspace.owner_email,
+                owner_first_name: workspace.owner_first_name,
+                owner_last_name: workspace.owner_last_name,
+                _owner_remote_id: workspace._owner,
+                environment: "PROD", // TODO
+                num_members: usersCount,
+                num_invited_users: workspace.invited_users.length,
+                num_groups: groupsCount,
+                created_date: workspace.created_date,
+                billing: {
+                    subscription_id: workspace.billing.subscription_id,
+                    current_period_end: workspace.billing.current_period_end,
+                    scheduled_cancellation: workspace.billing.scheduled_cancellation,
+                    quantity: usersCount
+                }
             }
+
+            http.put(`${process.env.MANAGEMENT_URL}/api/workspace/${workspace._id}/update`, {
+                API_KEY: process.env.MANAGEMENT_API_KEY,
+                workspaceData: workspaceMgmt
+            });
+
+            http.delete(`${process.env.MANAGEMENT_URL}/api/user/${userId}`, {
+                data: {
+                    API_KEY: process.env.MANAGEMENT_API_KEY
+                }
+            });
         }
-
-        http.put(`${process.env.MANAGEMENT_URL}/api/workspace/${workspace._id}/update`, {
-            API_KEY: process.env.MANAGEMENT_API_KEY,
-            workspaceData: workspaceMgmt
-        });
-
-        http.delete(`${process.env.MANAGEMENT_URL}/api/user/${userId}`, {
-            data: {
-                API_KEY: process.env.MANAGEMENT_API_KEY
-            }
-        });
 
         // Send the status 200 response 
         return res.status(200).json({
